requirements.txt
To run the projet you need to install all these librairies:

Pillow==10.0.1
pygame==2.5.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
tkinter


